OpenOCD: Reduce adapter speed.
The signals on the hardware are not in best shape. Reducing the speed a
bit makes the JTAG communication more reliable. Some tests showed that
the FTDI can't do any faster JTAG TCK then 6MHz. Any higher speed
will just lead to 6MHz. With that an example application loads with
41KB/sec. With 3MHz the same speed is reached. Some higher level
transfers seem to have more influence on the speed. At 1MHz, the
Speed drops to 26KB/sec. So 3MHz is a good selection with acceptable
signal quality.